# Implement integrity hashes