Use an [xxHash][1] crate to implement integrity hashes:

    xxHash is an Extremely fast Hash algorithm, running at RAM speed
    limits.

This required adding a hash field to Write, and a hashes field to
ReadResponse - a list of hashes must be returned to remain crash
consistent, just like with encryption context, and must be written
before extent data is.

Downstairs, a new table in sqlite was required, plus additional code for
unencrypted and encrypted branches. A new error was also created for
hash mismatches. Note that the downstairs can verify the hashes which
should aid in reconciliation.

Unfortunately adding integrity hashes caused import performance to drop
significantly. This is due to the fact that hashes have to be stored per
block, and the import code sends up to a 32M buffer at a time. New code
had to be added downstairs to split this up and hash each block during
the write operation. Rayon has been used downstairs to parallelize the
hash computation, but the big bottleneck was found to be the way I was
doing sqlite IMPORT WITH VALUES execution in set_hashes.

Performance was gained back by changing the large IMPORT WITH VALUES
from sending many values in a large statement to individual IMPORT
statements with one set of values each. This dramatically increased
performance, potentially pointing to some issue scaling query parsing or
planning in sqlite.

Importantly, extent_from_offset only supports single block address
output now. Guest reads and writes still can be any multiple of block
size but this is now always separated to single blocks when reading from
and writing to downstairs. Tests were updated to reflect this change.
This was done primarily to keep the downstairs simple, and to reflect
the decision that integrity hashes can be verified downstairs but should
not be generated downstairs. If a read of an arbitrary number of blocks
can be requested, the downstairs would have to compute the read response
hash instead of reading it from the database.

Integrity hashes were added to the dump output, along with dashes
separating table headers and rows:

    HASHES         0                1                2          DIFF
    ------  ---------------- ---------------- ---------------- -----
         0  16bdfe6adb2a4fe0 16bdfe6adb2a4fe1 16bdfe6adb2a4fe0  <---
         1  0011223344556677                                    <---

[1]: https://github.com/Cyan4973/xxHash