@jmpesp 2022-01-14
Copying performance results from #177 showing the degradation:
```
test:
f3write "${PWD}" && f3read "${PWD}"

before:
Average writing speed: 150.68 MB/s
Average reading speed: 384.96 MB/s

after:
Average writing speed: 11.86 MB/s (-170.81%)
Average reading speed: 28.93 MB/s (-172.04%)
```
```
test:
dd if=/dev/urandom of=/dev/nvme0n1 bs=512000 status=progress
dd if=/dev/nvme0n1 of=/dev/null bs=512000 status=progress

before:
write: 1073741824 bytes (1.1 GB, 1.0 GiB) copied, 22.6267 s, 47.5 MB/s
 read: 1073741824 bytes (1.1 GB, 1.0 GiB) copied, 6.31488 s, 170 MB/s

after:
write: 1073741824 bytes (1.1 GB, 1.0 GiB) copied, 62.3129 s, 17.2 MB/s (-93.66%)
 read: 1073741824 bytes (1.1 GB, 1.0 GiB) copied, 40.1809 s, 26.7 MB/s (-145.70%)
```