@jmpesp 2022-01-14
@leftwo this is failing because the integrity hashes are catching the fact that `./tools/test_up.sh` runs the client tests unencrypted, but the hammer test encrypted :) Unencrypted block writes store one integrity hash, then an encrypted write stores an encryption context + hash. You can reproduce this by creating an extent with only 1 block:

```
cargo run --release -q -p crucible-downstairs -- create -u 6d20be0f-49d6-447c-99ab-2e96a67cb834 -d disks/d1/ --extent-count 1 --extent-size 1 --block-size 4096
```

Then running the "one" workload once without an encryption key, then once with one. Dumping that block shows:
```
Extent 0 Block 0

  DATA  SHA256                                                            VER
------  ----------------------------------------------------------------  ---
     0  52a264cb5136333cbfaec4aecdf19d7f97b1d3a1d018ca07b89e47b898c1105c   A 
     1  52a264cb5136333cbfaec4aecdf19d7f97b1d3a1d018ca07b89e47b898c1105c   A 
     2  52a264cb5136333cbfaec4aecdf19d7f97b1d3a1d018ca07b89e47b898c1105c   A 

NONCES             0                        1                        2              DIFF 
------  ------------------------ ------------------------ ------------------------ ----- 
     0  10ac7cbbe77fdc9ed5a74e42 10ac7cbbe77fdc9ed5a74e42 10ac7cbbe77fdc9ed5a74e42 

  TAGS                 0                                1                                2                  DIFF 
------  -------------------------------- -------------------------------- -------------------------------- ----- 
     0  6cf97ea43dd8a747cdb346e74866cf14 6cf97ea43dd8a747cdb346e74866cf14 6cf97ea43dd8a747cdb346e74866cf14 

HASHES         0                1                2          DIFF 
------  ---------------- ---------------- ---------------- ----- 
     0  e08487ae5ca80665 e08487ae5ca80665 e08487ae5ca80665 
     1  124eed7401c2eda2 124eed7401c2eda2 124eed7401c2eda2 
```