@jmpesp 2022-01-14
The question is: do we want to support opening a downstairs with both unencrypted and encrypted Upstairs?