@jmpesp 2022-01-14
With

```
diff --git a/upstairs/src/lib.rs b/upstairs/src/lib.rs
index 6427ca1..ec8215a 100644
--- a/upstairs/src/lib.rs
+++ b/upstairs/src/lib.rs
@@ -1610,7 +1610,11 @@ impl Downstairs {
                             //
                             // XXX if it's not a blank block, we may be under
                             // attack?
-                            assert!(response.data[..].iter().all(|&x| x == 0));
+                            //
+                            // Note this can fire if an unencrypted write to the
+                            // block has occurred, then an encrypted read
+                            // occurs.
+                            //assert!(response.data[..].iter().all(|&x| x == 0));
                         }
                     }
 
@@ -1660,7 +1664,11 @@ impl Downstairs {
                             //
                             // XXX if it's not a blank block, we may be under
                             // attack?
-                            assert!(response.data[..].iter().all(|&x| x == 0));
+                            //
+                            // Note this can fire if an unencrypted write to the
+                            // block has occurred, then an encrypted read
+                            // occurs.
+                            //assert!(response.data[..].iter().all(|&x| x == 0));
                         }
                     }
 

```

test_up.sh's hammer run now fails with the same thing that the "one" workload reproduction does:
```
thread 'crucible-tokio' panicked at 'assertion failed: `(left == right)`
  left: `1`,
 right: `2`', upstairs/src/lib.rs:1556:29
note: run with `RUST_BACKTRACE=1` environment variable to display a backtrace
```