@leftwo 2022-01-14
> The question is: do we want to support opening a downstairs with both unencrypted and encrypted Upstairs?

No.  I don't think we want to, or can really.  If a block is written with encryption, then it should be encrypted forever.
I would not expect to switch back and forth between regions with encryption and regions without.

Maybe the test needs to be updated to create a fresh region for tests that are using keys.