@jmpesp 2022-01-17
@leftwo I added support for the downstairs to expect an unencrypted or an encrypted connection based on what it's seen before, see 5d0925c661ff23b205951219431ee021cfd2e9f9.