@jmpesp 2022-01-20
I'm pretty sure it's not broken anymore, yeah. Just ran hammer itself locally both unencrypted and encrypted and it worked.